modified:   mockDb/satDatabase.json 	deleted:    src/components/attendance/Attendance.css 	modified:   src/components/nav/NavBar.css 	modified:   src/components/nav/NavBar.js 	new file:   src/components/students/Attendance.css 	renamed:    src/components/attendance/Attendance.js -> src/components/students/Attendance.js 	new file:   src/components/students/Student.css 	new file:   src/components/students/Student.js 	modified:   src/components/views/ApplicationViews.js
