+++ mockDb/satDatabase.json
@@ -21,7 +21,7 @@
             "phone": "254-689147",
             "address": "1045 Bison Trl, Gallatin, TN 37066",
             "gradeId": 1,
-            "imgLink": ""
+            "imgLink": "https://media.istockphoto.com/id/1333890583/photo/photo-of-cute-charming-schoolboy-wear-white-shirt-spectacles-smiling-isolated-blue-color.jpg?s=612x612&w=0&k=20&c=FMAeWhMeNsUfVA21UEk8ZUtM6zm58FdYU9ZKWBgQAoI="
         },
         {
             "id": 2,
@@ -34,7 +34,7 @@
             "phone": "615-6024498",
             "address": "5017 Market Pl #5017, Mt. Juliet, TN 37122",
             "gradeId":1,
-            "imgLink": ""
+            "imgLink": "https://media.istockphoto.com/id/1055560054/photo/handsome-hispanic-man-model-wearing-winter-sweater-over-isolated-background-with-a-happy-and.jpg?s=612x612&w=0&k=20&c=uFydigVqffWhslBrv8qNLHJ-c84XkI9WlwHK09K6Yss="
         },
         {
             "id": 3,
@@ -103,8 +103,8 @@
         },
         {
             "id": 8,
-            "fullName": "eman Mena",
-            "email": "emena@me.com",
+            "fullName": "Thomas Samy",
+            "email": "tsamy@me.com",
             "userTypeId": 2,
             "userId": 8,
             "classId": 1,
