--- src/components/attendance/Attendance.css
@@ -1,31 +1,0 @@
-.attendance{
-    display: flex;
-    flex-direction: column;
-} 
-.attendanceM{
-    /* background-color:rgb(219, 235, 248); */
-    background-color:pink;
-    border:1px dotted rgb(133, 42, 54);
-    padding: 0.7rem;
-    margin: 0.5rem;
-    max-width:20rem
-}
-
-.student_card_container
-{
-    display: flex;
-    flex-direction: row;
-}
-li{
-    list-style-type: none;
-}
-.student_img{
-    width: 70px;
-    height: 70px;
-    /* background-size: cover; */
-    object-fit: cover; 
-    border-radius: 50%;
-}
-.student_info{
-    
-}
