+++ src/components/nav/NavBar.css
@@ -1,15 +1,84 @@
 .navbar {
     display: flex;
     flex-wrap: nowrap;
-    justify-content: left;
+    /* justify-content: left; */
+    justify-content: space-between;
 }
 
-.navbar__item {
-    flex-basis: 20%;
+.navbar_item {
+    flex-basis: 50%;
     list-style-type: none;
 }
-
+.navBar_link{
+margin: 10px;
+}
 .navbar__logout {
     margin-left: auto;
 }
 
+ li{
+    list-style-type: none;
+} 
+/* 
+@import url("https://fonts.googleapis.com/css?family=Open+Sans:400,300,600,700|Libre+Baskerville|News+Cycle");
+
+.navBar {
+    display: flex;
+    justify-content: center;
+    padding: 1vh 0;
+    list-style: none;
+    text-align: right;
+    font-family: "Open Sans";
+    text-transform: uppercase;
+    position: fixed;
+    width: 100%;
+    height: 120px;
+    top: -16px;
+    background-color: white;
+    z-index: 99;
+    border-bottom: #00b189 solid 3px;
+}
+
+.navBar ul {
+    list-style: none;
+    display: flex;
+    justify-content: space-evenly;
+    align-items: flex-end;
+    flex-wrap: wrap;
+    position: relative;
+    bottom: 10px;
+    padding-top: 2vh;
+}
+
+.navBar li {
+    width: 150px;
+    text-align: center;
+    font-size: 16px;
+}
+
+.navBar a {
+    text-decoration: none;
+    color: #2d1a02;
+    padding: 0 3vh;
+}
+
+.logo {
+    margin: 1vh;
+    padding-right: 5vh;
+    position: relative;
+    top: 0.9vh;
+    object-fit: cover;
+    object-position: center;
+}
+
+.navSpacer {
+    margin-bottom: 14vh;
+}
+
+@media only screen and (max-width: 1450px) {
+    .navBar li {
+        width: 115px;
+        font-size: 12px;
+    }
+}
+ */
