+++ src/components/nav/NavBar.js
@@ -4,11 +4,13 @@ import "./NavBar.css"
 export const NavBar = () => {
     const navigate = useNavigate()
 
-    return (
+    return(<>
+        <nav>
         <ul className="navbar">
             
             
         <li className="navBar_item"><Link className="navBar_link" to="/attendance">Attendance</Link></li>
+        <li className="navBar_item"><Link className="navBar_link" to="/">Home</Link></li>
     {/*     <li className="navBar_item"><Link className="navBar_link" to="/addNewStudent">Add New Student</Link></li>
         <li className="navBar_item"><Link className="navBar_link" to="/contacts">Contacts</Link></li>
         <li className="navBar_item"><Link className="navBar_link" to="/birthdays">BirthDays</Link></li>
@@ -25,7 +27,9 @@ export const NavBar = () => {
                 }}>Logout</Link>
             </li>
 
-        </ul>
+            </ul>
+            </nav>
+        </>
     )
 }
