+++ src/components/students/Attendance.css
@@ -1,0 +1,49 @@
+h1{
+    font-weight: lighter;
+    font-family:Verdana, Geneva, Tahoma, sans-serif;
+    color:rgb(30, 68, 121)
+}
+
+.attendance {
+    display: flex;
+    flex-direction: column;
+}
+.attendance_card {
+    /* background-color:rgb(219, 235, 248); */
+    background-color: rgba(255, 192, 203, 0.493);
+    border: 1px dotted rgb(133, 42, 54);
+    border-radius: 15px;
+    padding: 0.5rem;
+    margin: 0.3rem;
+    max-width: 19rem;
+    box-shadow: 1px 1px 3px rgb(196, 196, 196);
+}
+
+.student_card_container {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    align-items: center;
+}
+
+.student_img {
+    width: 70px;
+    height: 70px;
+    /* background-size: cover; */
+    object-fit: cover;
+    border-radius: 50%;
+}
+.student_info {
+    font-size: 1.2rem;
+    font-weight: lighter;
+    font-family: "Gill Sans", "Gill Sans MT", Calibri, "Trebuchet MS",
+        sans-serif;
+    color: rgb(5, 96, 96);
+}
+
+.checkmark {
+    margin: 1rem;
+    height: 20px;
+    width: 20px;
+}
+
