+++ src/components/students/Attendance.js
@@ -1,4 +1,5 @@
 import { useEffect, useState } from "react";
+import { Student } from "./Student";
 import "./Attendance.css";
 
 export const Attendance = () => {
@@ -9,7 +10,7 @@ export const Attendance = () => {
             const fetchData = async () => {
                 const response = await fetch(
                     `http://localhost:8033/users?userTypeId=2&classId=1`
-                    `http://localhost:8033/classes/1?_embed=users`
+                    //`http://localhost:8033/classes/1?_embed=users`
                 );
                 const attendanceArray = await response.json();
                 setAttendance(attendanceArray);
@@ -22,11 +23,26 @@ export const Attendance = () => {
     return (
         <>
             <h1>Attendance</h1>
-            {/* 
-            <p>
-                This is a simple example of how to use the{" "}
-                <code>Attendance</code> component.
-            </p> */}
+            <article className="attendance">
+                {attendance.map((student) => {
+                    return (
+                        <>
+                            <Student key={`ticket--${student.id}`}
+                                studentImg={student.imgLink}
+                                 studentId={student.id}
+                                 studentName={student.fullName}
+                            />
+                        </>
+                    );
+                })}
+            </article>
+        </>
+    );
+};
+
+/*     return (
+        <>
+            <h1>Attendance</h1>
             <article className="attendance">
                 {attendance.map((student) => {
                     return (
@@ -46,4 +62,4 @@ export const Attendance = () => {
             </article>
         </>
     );
-};
+}; */
